[ADD] Basic behavior tree and senses to AI agents
The AI can now perform an idle behavior by randomly choosing a point in a radius around her using a navigation agent.
The senses can see and store 1 enemy in the AI memory and if in range the AI will stop his idle stance and look directly to the enemy until he exit her range.